move method and type
