+++ src/app/pipes/resolution-filter.service.ts
@@ -2,8 +2,6 @@ import { Injectable } from '@angular/core';
 
 import { BehaviorSubject } from 'rxjs';
 
-export type ResolutionString = '' | 'SD' | '720' | '720+' | '1080' | '1080+' | '4K' | '4K+';
-
 @Injectable()
 export class ResolutionFilterService {
